+++ MyStr.h
@@ -20,7 +20,7 @@ int  utf8ToUnicode ( const char* lpMultiByteStr, WCHAR* lpWideCharStr, int cmbCh
 
 CString WtoUTF8(LPCTSTR str);
 CString UTF8toW(LPCTSTR str);
-CString FormatNumber(int str);
+CString FormatNumber(UINT64 str);
 
 void DelEndSlash(CString & str, bool bRev = false);
 void AddEndSlash(CString & str, bool bRev = false);
